requirements.txt
# Core trading dependencies
kiteconnect==4.2.0
pandas>=1.5.0
numpy>=1.24.0

# Web server for health checks
Flask>=2.3.0
gunicorn>=21.0.0

# Database
# sqlite3  # Built into Python (commented for cloud)

# Configuration
PyYAML>=6.0

# Telegram notifications
python-telegram-bot>=20.0

# Async programming
# asyncio  # Built into Python (commented)
aiohttp>=3.8.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Logging and monitoring
colorlog>=6.7.0

# Data validation
pydantic>=2.0.0

# Optional: For enhanced error handling
tenacity>=8.2.0

# Market data APIs
requests>=2.31.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
